refactor(codegen): rename gen-react/gen-vue to gen-react-19/gen-vue-3
