+++ scripts/codegen/src/generators/gen-react-19/_shared/branches.ts
