+++ scripts/codegen/src/generators/gen-react-19/kinds/atomic.ts
