+++ scripts/codegen/src/generators/gen-react-19/workspace/readme.ts
