+++ scripts/codegen/src/generators/gen-react.ts
@@ -11,13 +11,13 @@ import type { GeneratorContext, GeneratorReport } from "../registry.ts";
 import { registerGenerator } from "../registry.ts";
 import { Spec as SpecSchema } from "../schema.ts";
 import type { Spec } from "./gen-contract.ts";
-import { renderCssShim } from "./gen-react/_shared/css-shim.ts";
-import { renderAtomicReactWrapper } from "./gen-react/kinds/atomic.ts";
-import { renderCompositeListReactWrapper } from "./gen-react/kinds/composite-list.ts";
-import { renderCompositeOverlayReactWrapper } from "./gen-react/kinds/composite-overlay.ts";
-import { renderBarrel } from "./gen-react/workspace/barrel.ts";
-import { renderReadme } from "./gen-react/workspace/readme.ts";
-import { renderRuntime } from "./gen-react/workspace/runtime.ts";
+import { renderCssShim } from "./gen-react-19/_shared/css-shim.ts";
+import { renderAtomicReactWrapper } from "./gen-react-19/kinds/atomic.ts";
+import { renderCompositeListReactWrapper } from "./gen-react-19/kinds/composite-list.ts";
+import { renderCompositeOverlayReactWrapper } from "./gen-react-19/kinds/composite-overlay.ts";
+import { renderBarrel } from "./gen-react-19/workspace/barrel.ts";
+import { renderReadme } from "./gen-react-19/workspace/readme.ts";
+import { renderRuntime } from "./gen-react-19/workspace/runtime.ts";
 
 const REPO_ROOT = resolve(dirname(fileURLToPath(import.meta.url)), "..", "..", "..", "..");
 const SPECS_DIR = resolve(REPO_ROOT, "specs");
