+++ scripts/codegen/src/generators/gen-vue-3/_shared/props.test.ts
