+++ scripts/codegen/src/generators/gen-vue-3/_shared/type-printer.test.ts
