+++ scripts/codegen/src/generators/gen-vue-3/kinds/atomic.test.ts
