+++ scripts/codegen/src/generators/gen-vue-3/kinds/composite-list.test.ts
