+++ scripts/codegen/src/generators/gen-vue-3/kinds/composite-overlay.ts
