+++ scripts/codegen/src/generators/gen-vue-3/workspace/barrel.ts
