+++ scripts/codegen/src/generators/gen-vue.ts
@@ -12,13 +12,13 @@ import type { GeneratorContext, GeneratorReport } from "../registry.ts";
 import { registerGenerator } from "../registry.ts";
 import { Spec as SpecSchema } from "../schema.ts";
 import type { Spec } from "./gen-contract.ts";
-import { renderCssShim } from "./gen-vue/_shared/css-shim.ts";
-import { renderAtomicVueWrapper } from "./gen-vue/kinds/atomic.ts";
-import { renderCompositeListVueWrapper } from "./gen-vue/kinds/composite-list.ts";
-import { renderCompositeOverlayVueWrapper } from "./gen-vue/kinds/composite-overlay.ts";
-import { renderBarrel } from "./gen-vue/workspace/barrel.ts";
-import { renderReadme } from "./gen-vue/workspace/readme.ts";
-import { renderRuntime } from "./gen-vue/workspace/runtime.ts";
+import { renderCssShim } from "./gen-vue-3/_shared/css-shim.ts";
+import { renderAtomicVueWrapper } from "./gen-vue-3/kinds/atomic.ts";
+import { renderCompositeListVueWrapper } from "./gen-vue-3/kinds/composite-list.ts";
+import { renderCompositeOverlayVueWrapper } from "./gen-vue-3/kinds/composite-overlay.ts";
+import { renderBarrel } from "./gen-vue-3/workspace/barrel.ts";
+import { renderReadme } from "./gen-vue-3/workspace/readme.ts";
+import { renderRuntime } from "./gen-vue-3/workspace/runtime.ts";
 
 const REPO_ROOT = resolve(dirname(fileURLToPath(import.meta.url)), "..", "..", "..", "..");
 const SPECS_DIR = resolve(REPO_ROOT, "specs");
